[project @ SwkWidget.java change getBorderWidth to return int]

+++ src/com/onemoonscientific/swank/SwkWidget.java
@@ -63,7 +63,7 @@ public interface SwkWidget {
 
     public void setCreated(boolean state);
 
-    public double getBorderWidth();
+    public int getBorderWidth();
 
     public Insets getEmptyBorderInsets();
